refactor(db): extract checkpoint allocation ownership
